LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/keyframe-effects/effect-value-context-filling-expected.txt


FAIL Filling effect values reflect changes to font-size on element assert_equals: Effect value after updating font-size expected "400px" but got "200px"
FAIL Filling effect values reflect changes to font-size on parent element assert_equals: Effect value after updating font-size on parent element expected "400px" but got "200px"
FAIL Filling effect values reflect changes to variables on element assert_equals: Effect value after updating variable expected "200px" but got "100px"
FAIL Filling effect values reflect changes to variables on parent element assert_equals: Effect value after updating variable expected "400px" but got "200px"
FAIL Filling effect values reflect changes to the the animation's keyframes assert_equals: Effect value after updating the animation expected "300px" but got "200px"
FAIL Filling effect values reflect changes to the the animation's composite mode assert_equals: Effect value after updating the composite mode expected "300px" but got "200px"
FAIL Filling effect values reflect changes to the the animation's iteration composite mode assert_equals: Effect value after updating the iteration composite mode expected "200px" but got "100px"
FAIL Filling effect values reflect changes to the base value when using additive animation assert_equals: Effect value before updating underlying value expected "300px" but got "200px"
FAIL Filling effect values reflect changes to the base value when using additive animation on a single keyframe assert_equals: Effect value before updating underlying value expected "300px" but got "200px"
PASS Filling effect values reflect changes to the base value when using the fill value is an implicit keyframe
FAIL Filling effect values reflect changes to the base value via a parent element assert_equals: Effect value before updating font-size on parent expected "300px" but got "200px"
FAIL Filling effect values reflect changes to underlying animations assert_equals: Effect value after updating underlying animation expected "300px" but got "200px"
FAIL Filling effect values reflect changes to underlying animations via a a parent element assert_equals: Effect value before updating parent font-size expected "300px" but got "200px"
FAIL Filling effect values reflect changes to underlying animations made by directly changing the keyframes assert_equals: Effect value after updating underlying animation expected "300px" but got "200px"
